Go_101: Completed Program 18: Solid Box of Asterisks.
Was fun to do, and reinforced my learning on the different forms
of a for loop. I have yet to mention the third one, which is
just a loop without any conditions, hence running until code in
the block specifies breaking out of it. I did not get to use a
range clause, will keep it in my arsenal to use later...